Baltimore City and Regional Plans

Table from the Baltimore City and Regional Plans Airtable base: https://airtable.com/appC0OOaNJtKU6iHz/shrIb0577rFcNI6vj

Usage

baltimore_plans

Format

A data frame with 366 rows and 36 variables:

title

Plan title

year_complete

Year adopted/completed

year_update

Year updated

date_complete

Date (may be approximate)

status

Current status

document_url

Document URL

document_filename

Document URL

document_cover_filename

Document URL

site_url

Related site URL

geography

Geography

document_type

Document type

program

Related program

prepared_for

Entity or entities prepared for

prepared_by

Entity or entities prepared by

funded_by

Entity or entities funded by

location_names

Location names

county

County name

main_topic

Main topic

topics

Topics

name_abb

Name abbreviation

name_short

Short name

title_alt

Alternate title for spatial data join

short_description

Short document description

document_page_count

Page count

document_url_type

Page count

status_organization

Entity assigning status

year_status

Status assignment date

date_status

Status assignment date

status_regulatory

Regulatory status

source

Document source

related_plans

Related plans (typically a preceding plan in a sequence of related plans)

notes

Notes

created_date

Airtable record created date/time

last_modified

Airtable record last modified date/time

record_id

Airtable record identifier

Source

https://airtable.com/appC0OOaNJtKU6iHz/shrIb0577rFcNI6vj

Details

This base is an index for community plans, park plans, transportation plans, and related reports or studies for Baltimore City, adjacent counties, and the Baltimore metro area.

The goal is to provide a convenient and accessible reference to information plans include: what the scope and related locations for the plan, when the plan was published or adopted, who prepared the plan, and any programs connected to the plan. This data is the plans table including plans, studies, reports, etc.

Last updated: 2025-05-22