BNIA Vital Signs Indicators, 2010-2018

Labels, definitions, and data sources for the Vital Signs indicators used by the Baltimore Neighborhood Indicators Alliance from 2010 to 2018.

vital_signs_indicators

Format

A data frame with 177 rows and 7 variables:

label

Indicator label

section

Vital Signs report section

source

Indicator data source

normalization_source

Normalization data source, e.g. Census data for CSA residents

definition

Indicator definition

indicator

Indicator variable name

years_available

Years available

Source

https://bniajfi.org/vital_signs/