BNIA Vital Signs, 2010-2018

The Baltimore Neighborhood Indicator Alliance (BNIA) describes the Vital Signs are "data-points compiled from a variety of reliable sources that 'take the pulse' of Baltimore’s neighborhoods." This data set is based on the Vital Signs Data Tables (2010-2018) master Excel file containing all Vital Signs indicators, by year of release.

Usage

vital_signs_2010_2018

Format

A data frame with 62552 rows and 9 variables:

csa2010

Community Statistical Area name

year

Vital Signs year

indicator

Indicator name

value

Value of indicator

label

Indicator label

section

Vital Signs section, matching one of eight topic areas

Source

https://bniajfi.org/vital_signs/

Details

The source file contained the full name of the indicator, values for Baltimore City and all 55 of the Community Statistical Areas (CSAs). Due to COVID-19, the publication of the Vital Signs 18 indicators has been staggered and this data does not yet contain full housing or education data. CSAs with missing data in the Excel file appear as NA values in this dataset.

See also